Exceptions Handeled

+++ src/Main.java
@@ -85,14 +85,32 @@ public class Main {
                         System.out.println("Account name: ");
                         String name = scanner.next();
 
-                        service.addACustomer(email, name);
+                        try
+                        {
+                            service.checkForEmail(email);
+                            service.addACustomer(email, name);
+                        }
+                        catch (RuntimeException e)
+                        {
+                            System.out.println("There is already an account created with this email!");
+                        }
                     }
                     else
                     {
-                        System.out.println("Introduce email to delete: ");
+                        System.out.println("Introduce your email to delete: ");
                         String email = scanner.next();
 
-                        service.removeACustomer(email);
+                        try
+                        {
+                            if(service.testEmail(service.accountId, email)==5)
+                            {
+                                service.removeACustomer(email);
+                            }
+
+                        }catch (RuntimeException e)
+                        {
+                            System.out.println("The email you introduced was wrong!");
+                        }
                     }
                     break;
 
@@ -121,7 +139,21 @@ public class Main {
                         System.out.println("Introduce the sum you would like to deposit or withdraw: ");
                         String dep = scanner.next();
 
-                        service.addOrWithdraw(depwith, dep);
+                        try
+                        {
+                            if(Integer.parseInt(dep)>=0)
+                            {
+                                service.addOrWithdraw(depwith, dep);
+                            }
+                            else
+                            {
+                                System.out.println("You cannot introduce negative balance!");
+                            }
+                        }
+                        catch (RuntimeException e)
+                        {
+                            System.out.println("Invalid sum!");
+                        }
                     }
                     else
                     {
@@ -144,52 +176,64 @@ public class Main {
                             }
                             else
                             {
-                                System.out.println("Please choose a collection name: ");
-                                String collectionName = scanner.next();
+                                try
+                                {
+                                    System.out.println("Please choose a collection name: ");
+                                    String collectionName = scanner.next();
 
-                                System.out.println("Please choose a machine gun skin: ");
-                                String machineGunSkin = scanner.next();
+                                    System.out.println("Please choose a machine gun skin: ");
+                                    String machineGunSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the machine gun skin: ");
-                                float floatCapMachineGunSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the machine gun skin: ");
+                                    float floatCapMachineGunSkin = Float.parseFloat(scanner.next());
 
-                                System.out.println("Please choose a shotgun skin: ");
-                                String shotgunSkin = scanner.next();
+                                    System.out.println("Please choose a shotgun skin: ");
+                                    String shotgunSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the shotgun skin: ");
-                                float floatCapShotgunSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the shotgun skin: ");
+                                    float floatCapShotgunSkin = Float.parseFloat(scanner.next());
 
-                                System.out.println("Please choose a smg skin: ");
-                                String smgSkin = scanner.next();
+                                    System.out.println("Please choose a smg skin: ");
+                                    String smgSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the smg skin: ");
-                                float floatCapSmgSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the smg skin: ");
+                                    float floatCapSmgSkin = Float.parseFloat(scanner.next());
 
-                                System.out.println("Please choose a pistol skin: ");
-                                String pistolSkin = scanner.next();
+                                    System.out.println("Please choose a pistol skin: ");
+                                    String pistolSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the pistol skin: ");
-                                float floatCapPistolSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the pistol skin: ");
+                                    float floatCapPistolSkin = Float.parseFloat(scanner.next());
 
-                                System.out.println("Please choose a sniper rifle skin: ");
-                                String sniperRifleSkin = scanner.next();
+                                    System.out.println("Please choose a sniper rifle skin: ");
+                                    String sniperRifleSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the sniper rifle skin: ");
-                                float floatCapSniperRifleSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the sniper rifle skin: ");
+                                    float floatCapSniperRifleSkin = Float.parseFloat(scanner.next());
 
-                                System.out.println("Please choose a rifle skin: ");
-                                String rifleSkin = scanner.next();
+                                    System.out.println("Please choose a rifle skin: ");
+                                    String rifleSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the rifle skin: ");
-                                float floatCapRifleSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the rifle skin: ");
+                                    float floatCapRifleSkin = Float.parseFloat(scanner.next());
 
-                                System.out.println("Please choose a knife skin: ");
-                                String knifeSkin = scanner.next();
+                                    System.out.println("Please choose a knife skin: ");
+                                    String knifeSkin = scanner.next();
 
-                                System.out.println("Please choose a float cap for the knife skin: ");
-                                float floatCapKnifeSkin = Float.parseFloat(scanner.next());
+                                    System.out.println("Please choose a float cap for the knife skin: ");
+                                    float floatCapKnifeSkin = Float.parseFloat(scanner.next());
 
-                                service.createASkinCollection(collectionName, machineGunSkin, floatCapMachineGunSkin, shotgunSkin, floatCapShotgunSkin, smgSkin, floatCapSmgSkin, pistolSkin, floatCapPistolSkin, sniperRifleSkin, floatCapSniperRifleSkin, rifleSkin, floatCapRifleSkin, knifeSkin, floatCapKnifeSkin);
+                                    if(floatCapMachineGunSkin < 0f || floatCapShotgunSkin < 0f || floatCapSmgSkin < 0f || floatCapPistolSkin < 0f || floatCapSniperRifleSkin < 0f || floatCapRifleSkin < 0f || floatCapKnifeSkin < 0f)
+                                    {
+                                        throw new RuntimeException();
+                                    }
+
+                                    service.createASkinCollection(collectionName, machineGunSkin, floatCapMachineGunSkin, shotgunSkin, floatCapShotgunSkin, smgSkin, floatCapSmgSkin, pistolSkin, floatCapPistolSkin, sniperRifleSkin, floatCapSniperRifleSkin, rifleSkin, floatCapRifleSkin, knifeSkin, floatCapKnifeSkin);
+                                }
+                                catch (RuntimeException e)
+                                {
+                                    System.out.println("The float value introduced was not valid!");
+                                }
                             }
                         }
                     }else
@@ -213,40 +257,52 @@ public class Main {
                             }
                             else
                             {
-                                System.out.println("Please choose a graffiti set name: ");
-                                String graffitiSetName = scanner.next();
+                                try
+                                {
+                                    System.out.println("Please choose a graffiti set name: ");
+                                    String graffitiSetName = scanner.next();
 
-                                System.out.println("Please choose a support style graffiti: ");
-                                String support = scanner.next();
+                                    System.out.println("Please choose a support style graffiti: ");
+                                    String support = scanner.next();
 
-                                System.out.println("Please choose the number of uses for the support graffiti: ");
-                                Integer numberOfUsesSupport = Integer.parseInt(scanner.next());
+                                    System.out.println("Please choose the number of uses for the support graffiti: ");
+                                    Integer numberOfUsesSupport = Integer.parseInt(scanner.next());
 
-                                System.out.println("Please choose a leader style graffiti: ");
-                                String leader = scanner.next();
+                                    System.out.println("Please choose a leader style graffiti: ");
+                                    String leader = scanner.next();
 
-                                System.out.println("Please choose the number of uses for the leader graffiti: ");
-                                Integer numberOfUsesLeader = Integer.parseInt(scanner.next());
+                                    System.out.println("Please choose the number of uses for the leader graffiti: ");
+                                    Integer numberOfUsesLeader = Integer.parseInt(scanner.next());
 
-                                System.out.println("Please choose a ninja style graffiti: ");
-                                String ninja = scanner.next();
+                                    System.out.println("Please choose a ninja style graffiti: ");
+                                    String ninja = scanner.next();
 
-                                System.out.println("Please choose the number of uses for the ninja graffiti: ");
-                                Integer numberOfUsesNinja = Integer.parseInt(scanner.next());
+                                    System.out.println("Please choose the number of uses for the ninja graffiti: ");
+                                    Integer numberOfUsesNinja = Integer.parseInt(scanner.next());
 
-                                System.out.println("Please choose a fragger style graffiti: ");
-                                String fragger = scanner.next();
+                                    System.out.println("Please choose a fragger style graffiti: ");
+                                    String fragger = scanner.next();
 
-                                System.out.println("Please choose the number of uses for the fragger graffiti: ");
-                                Integer numberOfUsesFragger = Integer.parseInt(scanner.next());
+                                    System.out.println("Please choose the number of uses for the fragger graffiti: ");
+                                    Integer numberOfUsesFragger = Integer.parseInt(scanner.next());
 
-                                System.out.println("Please choose a bot style graffiti: ");
-                                String bot = scanner.next();
+                                    System.out.println("Please choose a bot style graffiti: ");
+                                    String bot = scanner.next();
 
-                                System.out.println("Please choose the number of uses for the bot graffiti: ");
-                                Integer numberOfUsesBot = Integer.parseInt(scanner.next());
+                                    System.out.println("Please choose the number of uses for the bot graffiti: ");
+                                    Integer numberOfUsesBot = Integer.parseInt(scanner.next());
 
-                                service.createAGraffitiSet(graffitiSetName, support, numberOfUsesSupport, leader, numberOfUsesLeader, ninja, numberOfUsesNinja, fragger, numberOfUsesFragger, bot, numberOfUsesBot);
+                                    if(numberOfUsesBot < 0 || numberOfUsesFragger < 0 || numberOfUsesLeader < 0 || numberOfUsesNinja < 0 || numberOfUsesSupport < 0)
+                                    {
+                                        throw new RuntimeException();
+                                    }
+
+                                    service.createAGraffitiSet(graffitiSetName, support, numberOfUsesSupport, leader, numberOfUsesLeader, ninja, numberOfUsesNinja, fragger, numberOfUsesFragger, bot, numberOfUsesBot);
+                                }
+                                catch (RuntimeException e)
+                                {
+                                    System.out.println("The number of uses introduced are invalid!");
+                                }
                             }
                         }
                     }else
@@ -271,9 +327,16 @@ public class Main {
                         if (conts > 0)
                         {
                             System.out.println("Those are your avalible containers. To open a container it costs 2. Please choose the id of the container you would like to open, or 0 to go back: ");
-                            Integer id = Integer.parseInt(scanner.next());
 
-                            service.accountOpenContainer(id);
+                            try
+                            {
+                                Integer id = Integer.parseInt(scanner.next());
+                                service.accountOpenContainer(id);
+                            }
+                            catch (RuntimeException e)
+                            {
+                                System.out.println("The container Id introduced is not valid!");
+                            }
                         }
                         else
                         {
@@ -291,16 +354,30 @@ public class Main {
                         System.out.println("Those are the items you have in your inventory!");
                         System.out.println("Please select an id of an item you would like to sell, or 0 to go back.");
 
-                        Integer id = Integer.parseInt(scanner.next());
-
-                        if (id != 0)
+                        try
                         {
-                            System.out.println("The creator of the collection will get 7% of the asking price, and the market will get 3%.");
-                            System.out.println("Choose a price for the item: ");
+                            Integer id = Integer.parseInt(scanner.next());
 
-                            Float pret = Float.parseFloat(scanner.next());
+                            if (id != 0)
+                            {
+                                System.out.println("The creator of the collection will get 7% of the asking price, and the market will get 3%.");
+                                System.out.println("Choose a price for the item: ");
 
-                            service.sellAccountProduct(id, pret);
+                                Float pret = Float.parseFloat(scanner.next());
+
+                                if(pret > 0f)
+                                {
+                                    service.sellAccountProduct(id, pret);
+                                }
+                                else
+                                {
+                                    System.out.println("The price cannot be negative!");
+                                }
+                            }
+                        }
+                        catch (RuntimeException e)
+                        {
+                            System.out.println("There has been a problem putting your item for sale! Check the item id and the price.");
                         }
                     }else
                     {
@@ -313,8 +390,14 @@ public class Main {
                         service.showMarketProducts();
                         System.out.println("Please choose the id of the item you would like to buy, or 0 to go back: ");
 
-                        Integer id = Integer.parseInt(scanner.next());
-                        service.buyTheItem(id);
+                        try{
+                            Integer id = Integer.parseInt(scanner.next());
+                            service.buyTheItem(id);
+                        }
+                        catch (RuntimeException e)
+                        {
+                            System.out.println("The id you introduced is invalid!");
+                        }
                     }else
                     {
                         System.out.println("Please create an account, or login!");
@@ -335,12 +418,13 @@ public class Main {
                     break;
 
                 default:
-                    System.out.println("Invalid command.");
+                    System.out.println("Command not recognized. Returning to main menu.");
                     break;
             }
         }   while (!command.equals("12"));
 
         scanner.close();
         System.out.println("Goodbye!");
+        System.out.println(service.profidMade());
     }
 }
